requirements.txt
asgiref==3.5.1
Django==4.0.4
django-ckeditor==6.4.2
django-js-asset==2.0.0
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2